requirements.txt
Django==3.1.1
mysqlclient==2.0.1
djangorestframework==3.11.0
requests==2.21.0